mymealplanner/requirements.txt
google-adk
opentelemetry-instrumentation-google-genai
flask
flask-cors
gunicorn
